remove some '#pragma comment lib', which are now handled by cmake
git-svn-id: http://vtp.googlecode.com/svn/trunk@7495 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/xmlhelper/easyxml.cpp
@@ -213,7 +213,7 @@ void readXML (istream &input, XMLVisitor &visitor, const string &path,
 		}
 
 		input.read(buf,16384);
-		if (!XML_Parse(parser, buf, input.gcount(), false))
+		if (!XML_Parse(parser, buf, (int) input.gcount(), false))
 		{
 			const XML_LChar *message = XML_ErrorString(XML_GetErrorCode(parser));
 			int line = XML_GetCurrentLineNumber(parser);
